tec: Refactor the code to find the authorized organizations
